[primitives] move bs58 stuff to its own module
